Created all main components They'll be called cards from now on.
